add continuous checks for zelapps hashes serviceManager
