+++ ZelBack/src/services/zelfluxCommunication.js
@@ -9,7 +9,6 @@ const log = require('../lib/log');
 const serviceHelper = require('./serviceHelper');
 const zelcashService = require('./zelcashService');
 const userconfig = require('../../../config/userconfig');
-const explorerService = require('./explorerService');
 
 const outgoingConnections = []; // websocket list
 const outgoingPeers = []; // array of objects containing ip and rtt latency
@@ -1186,52 +1185,6 @@ async function broadcastTemporaryZelAppMessage(message) {
   return 0;
 }
 
-async function startFluxFunctions() {
-  try {
-    log.info('Initiating MongoDB connection');
-    await serviceHelper.initiateDB(); // either true or throws error
-    log.info('DB connected');
-    log.info('Preparing local database...');
-    const db = serviceHelper.databaseConnection();
-    const database = db.db(config.database.local.database);
-    await serviceHelper.dropCollection(database, config.database.local.collections.activeLoginPhrases).catch((error) => {
-      if (error.message !== 'ns not found') {
-        log.error(error);
-      }
-    });
-    await serviceHelper.dropCollection(database, config.database.local.collections.activeSignatures).catch((error) => {
-      if (error.message !== 'ns not found') {
-        log.error(error);
-      }
-    });
-    await database.collection(config.database.local.collections.activeLoginPhrases).createIndex({ createdAt: 1 }, { expireAfterSeconds: 900 });
-    await database.collection(config.database.local.collections.activeSignatures).createIndex({ createdAt: 1 }, { expireAfterSeconds: 900 });
-    log.info('Local database prepared');
-    log.info('Preparing temporary database...');
-    // no need to drop temporary messages
-    const databaseTemp = db.db(config.database.zelappsglobal.database);
-    await databaseTemp.collection(config.database.zelappsglobal.collections.zelappsTemporaryMessages).createIndex({ createdAt: 1 }, { expireAfterSeconds: 3600 });
-    log.info('Temporary database prepared');
-    adjustFirewall();
-    fluxDisovery();
-    log.info('Flux Discovery started');
-    keepConnectionsAlive();
-    keepIncomingConnectionsAlive();
-    checkDeterministicNodesCollisions();
-    setInterval(() => {
-      checkDeterministicNodesCollisions();
-    }, 60000);
-    log.info('Flux checks operational');
-    explorerService.initiateBlockProcessor(true, true);
-    log.info('Flux Block Processing Service started');
-  } catch (e) {
-    log.error(e);
-    setTimeout(() => {
-      startFluxFunctions();
-    }, 15000);
-  }
-}
-
 module.exports = {
   getFluxMessageSignature,
   verifyOriginalFluxBroadcast,
@@ -1249,7 +1202,6 @@ module.exports = {
   serialiseAndSignZelFluxBroadcast,
   initiateAndHandleConnection,
   connectedPeers,
-  startFluxFunctions,
   addPeer,
   getIncomingConnections,
   getIncomingConnectionsInfo,
@@ -1265,4 +1217,8 @@ module.exports = {
   incomingPeers,
   isCommunicationEstablished,
   broadcastTemporaryZelAppMessage,
+  keepIncomingConnectionsAlive,
+  keepConnectionsAlive,
+  adjustFirewall,
+  checkDeterministicNodesCollisions,
 };
